Embed the manager.glade file and put it in a temporary file on start.
Unfortunately, the xmlNewWithRootAndDomain function really expects a FilePath to file, so we write the embedded glade file to a temporary file and delete it when the window is destroyed.

The deletion code is naive and ignores permission issues and other exceptions. In the event of trouble, we might want to update this to handle those.